backend/requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.0
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
PyJWT==2.10.1
python-dotenv==1.1.0
Werkzeug==3.1.3
Flask-Cors
